refactor: moved types/types.ts to types/question_paper.ts
